Move door handling to separate class
