+++ src/main.cpp
@@ -12,100 +12,14 @@
 #include "ota.h"
 #include <Arduino.h>
 
-#include <AccelStepper.h>
-#include <BH1750.h>
-#include <Wire.h>
-
 #include <ArduinoJson.h>
 
+#include "door.h"
 #include "mqtt-handler.h"
-
-#ifdef ESP32
-
-#define MOTOR_PIN1 GPIO_NUM_32
-#define MOTOR_PIN2 GPIO_NUM_33
-#define MOTOR_PIN3 GPIO_NUM_25
-#define MOTOR_PIN4 GPIO_NUM_26
-
-#define LIGHT_SDA GPIO_NUM_13
-#define LIGHT_SCL GPIO_NUM_15
-
-#define OPEN_PIN GPIO_NUM_14
-#define CLOSED_PIN GPIO_NUM_12
-
-#elif defined(ESP8266)
-
-#define OPEN_PIN D5
-#define CLOSED_PIN D6
-
-#define MOTOR_PIN1 D0
-#define MOTOR_PIN2 D1
-#define MOTOR_PIN3 D2
-#define MOTOR_PIN4 D3
-
-#define LIGHT_SDA D7
-#define LIGHT_SDC D4
-
-#endif
-
-const int stepsPerRevolution = 2048;
-
-AccelStepper motor(AccelStepper::FULL4WIRE, MOTOR_PIN1, MOTOR_PIN3, MOTOR_PIN2, MOTOR_PIN4);
-
-BH1750 lightMeter;
 
 Ota ota;
 
-struct Config {
-    /**
-     * Light required to be above limit to open the door.
-     */
-    float openLightLimit = 120;
-
-    /**
-     * Light required to be below limit to open the door.
-     */
-    float closeLightLimit = 50;
-
-    /**
-     * Whether to invert the "gate open" switch or not.
-     */
-    bool invertOpenSwitch = true;
-
-    /**
-     * Whether to invert the "gate close" switch or not.
-     */
-    bool invertCloseSwitch = true;
-} config;
-
-enum class GateState {
-    OPEN,
-    CLOSED,
-    OPENING,
-    CLOSING
-};
-
-struct State {
-    /**
-     * The current level of light.
-     */
-    float currentLight = 0;
-
-    /**
-     * The state of the gate.
-     */
-    GateState gateState = GateState::OPEN;
-
-    /**
-     * Whether the "gate open" switch is engaged or not.
-     */
-    bool openSwitch = false;
-
-    /**
-     * Whether the "gate closed" switch is engaged or not.
-     */
-    bool closedSwitch = false;
-} state;
+Door door;
 
 void LOG(const char* message) {
     Serial.println(message);
@@ -115,8 +29,6 @@ void setup() {
     Serial.begin(115200);
 
     pinMode(LED_BUILTIN, OUTPUT);
-    pinMode(OPEN_PIN, INPUT_PULLUP);
-    pinMode(CLOSED_PIN, INPUT_PULLUP);
 
     while (!Serial) {
         delay(100);
@@ -144,18 +56,6 @@ void setup() {
     LittleFS.setConfig(cfg);
     LittleFS.begin();
 #endif
-
-    motor.setMaxSpeed(500);
-    motor.setSpeed(500);
-    motor.setAcceleration(500);
-
-    Wire.begin(LIGHT_SDA, LIGHT_SCL);
-    Serial.println("Connecting to light sensor...");
-    if (lightMeter.begin()) {
-        Serial.println("BH1750 initialised");
-    } else {
-        Serial.println("Error initialising BH1750");
-    }
 
     WiFi.mode(WIFI_AP_STA);
     delay(500);
@@ -201,56 +101,12 @@ void setup() {
     WiFiClientSecure* wifiClient = new WiFiClientSecure();
     wifiClient->setCACert(root_cert.c_str());
     mqttHandler.begin(wifiClient, iotConfigJson);
+
+    door.begin();
 }
-
-unsigned long previousMillis = 0;
-unsigned long interval = 1000;
-
-unsigned int stepsAtOnce = 100;
 
 void loop() {
     ota.handle();
     mqttHandler.loop();
-
-    state.openSwitch = digitalRead(OPEN_PIN) ^ config.invertOpenSwitch;
-    state.closedSwitch = digitalRead(CLOSED_PIN) ^ config.invertCloseSwitch;
-
-    unsigned long currentMillis = millis();
-    if (currentMillis - previousMillis > interval) {
-        previousMillis = currentMillis;
-        state.currentLight = lightMeter.readLightLevel();
-        if (state.currentLight < config.closeLightLimit && state.gateState == GateState::OPEN) {
-            LOG("Closing...");
-            state.gateState = GateState::CLOSING;
-        } else if (state.currentLight > config.openLightLimit && state.gateState == GateState::CLOSED) {
-            LOG("Opening...");
-            state.gateState = GateState::OPENING;
-        }
-    }
-
-    if (!motor.run()) {
-        if (state.gateState == GateState::OPEN || state.gateState == GateState::CLOSED) {
-            motor.disableOutputs();
-            delay(250);
-            return;
-        }
-    }
-
-    if (state.gateState == GateState::CLOSING) {
-        if (state.closedSwitch) {
-            LOG("Closed");
-            motor.stop();
-            state.gateState = GateState::CLOSED;
-        } else {
-            motor.move(stepsAtOnce);
-        }
-    } else if (state.gateState == GateState::OPENING) {
-        if (state.openSwitch) {
-            LOG("Open");
-            motor.stop();
-            state.gateState = GateState::OPEN;
-        } else {
-            motor.move(-stepsAtOnce);
-        }
-    }
+    door.loop();
 }
